Back-End/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
awscli==1.22.34
bcrypt==4.2.1
blinker==1.4
botocore==1.23.34
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.1.8
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.17.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.7
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.5
jmespath==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.8
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
orjson==3.10.15
packaging==24.2
passlib==1.7.4
Pillow==9.0.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==5.4.1
requests==2.25.1
rich==13.9.4
rich-toolkit==0.13.2
roman==3.3
rsa==4.8
s3transfer==0.5.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
systemd-python==234
tomli==2.2.1
ttystatus==0.38
typer==0.15.1
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.34.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.0.4
websockets==14.2
zipp==1.0.0
